Move optional imports to imports_helper.py
