+++ python/seldon_core/proto/tensorflow/core/framework/__init__.py
@@ -1,18 +1,5 @@
-import logging
-import textwrap
+from seldon_core.imports_helper import _TF_MISSING
 
-logger = logging.getLogger(__name__)
-
-try:
+if not _TF_MISSING:
     # Let tensorflow shadow these imports if present
-    from tensorflow.core.framework import (
-        tensor_pb2, types_pb2, tensor_shape_pb2)
-except ImportError:
-    notice = textwrap.dedent("""
-        Tensorflow is not installed.
-        If you want to use `tftensor` and Tensorflow's data types
-        install `tensorflow` or install `seldon_core` as
-
-            $ pip install seldon_core[tensorflow]
-    """)
-    logger.info(notice)
+    from tensorflow.core.framework import tensor_pb2, types_pb2, tensor_shape_pb2
