--- python/seldon_core/tf_helper.py
@@ -1,9 +1,0 @@
-# Variable to check if TF is present or not
-_TF_MISSING = True
-
-try:
-    import tensorflow  # noqa: F401
-
-    _TF_MISSING = False
-except ImportError:
-    _TF_MISSING = True
