+++ python/seldon_core/utils.py
@@ -1,5 +1,4 @@
 import json
-import logging
 import sys
 import base64
 import numpy as np
@@ -17,7 +16,7 @@ from seldon_core.user_model import (
     client_feature_names,
     SeldonComponent,
 )
-from seldon_core.tf_helper import _TF_MISSING
+from seldon_core.imports_helper import _TF_MISSING
 
 from typing import Tuple, Dict, Union, List, Optional, Iterable
