+++ python/tests/test_utils.py
@@ -6,7 +6,7 @@ import base64
 import seldon_core.utils as scu
 from seldon_core.proto import prediction_pb2
 from seldon_core.flask_utils import SeldonMicroserviceException
-from seldon_core.tf_helper import _TF_MISSING
+from seldon_core.imports_helper import _TF_MISSING
 from google.protobuf.struct_pb2 import Value
 from utils import skipif_tf_missing
